+++ noodles-bam/src/lazy/record/read_name.rs
@@ -10,6 +10,14 @@ impl<'a> ReadName<'a> {
     pub(super) fn new(src: &'a [u8]) -> Self {
         Self(src)
     }
+
+    /// Returns the read name as a byte slice.
+    ///
+    /// The returned slice will _not_ have the trailing `NUL` terminator.
+    pub fn as_bytes(&self) -> &[u8] {
+        const NUL: u8 = 0x00;
+        self.as_ref().strip_suffix(&[NUL]).unwrap_or(self.as_ref())
+    }
 }
 
 impl<'a> AsRef<[u8]> for ReadName<'a> {
@@ -33,5 +41,19 @@ impl<'a> TryFrom<ReadName<'a>> for sam::record::ReadName {
             .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))?;
 
         Ok(read_name.unwrap())
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_as_bytes() {
+        let read_name = ReadName::new(b"r0\x00");
+        assert_eq!(read_name.as_bytes(), b"r0");
+
+        let read_name = ReadName::new(b"r0");
+        assert_eq!(read_name.as_bytes(), b"r0");
     }
 }