+++ noodles-bam/CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### Added
 
+  * bam/lazy/record/read_name: Add `ReadName::as_bytes` to return the buffer
+    without the trailing `NUL` terminator.
+
   * bam/lazy/record/sequence: Add bases iterator (`Sequence::iter`).
 
 ## 0.46.0 - 2023-09-21